Correction of the bug of landmark modification by user: try 1
When the user is invited to possibly modify some landmarks once all are
placed, a bug appear sometimes  which made that the point selectionned
by the user is not correct and another point is selected, with a change
in the numerotation. At the end th programm craches...

The correction here (the transfer of the two lines Idx<<-... and IdxPts
inside the if(firsTime)) seems to avoid the change of point labelling
during correction.
=> Need to be confirmed by users...