Correction of the bug of the non respected integrity of submesh in PlacePt before the use of vcgIsolated: try 1
When the user isi invited to place landmarks, a bug can occur during the
placement which crashes the main function. It seems to be due to the use
of vcgIsolated in PlacePt() which requires mesh integrity otherwise this
function stops and gives an error. Then it implies a crash of
DigitFixes.
=> A check of mesh integrity for SelecMesh was then added before the use
of vcgIsolated to prevent this bug... If not respected, none point is
selected, and the user can continue to place landmark.

Additionnaly: the two calls of vcvIsolated (in PlacePt() and in
SetPtZoom()) were set to silent to avoid the multiple lines "mesh is
split into n pieces" printed in the console...