Correction on the linking of intersection points from mesh and orthogonal planes.
The previous version of the function linked the points of intersection
(when asked) between the mesh and the 3 orthoplanes in an uncorrect way.
Indeed, a gross approximation was used for Catherine's trilobite using a
sorting of angles made between vectors from mesh centroids to all
intersection points. It roughly worked because of the globular shape of
the meshes. But this approximation didn't work for more complexe shapes.

The core correction here consists in the writing of the new function
sortCurveMesh (written in Rccp to faster its execution) that determines
in which order the intersection points should be linked for plotting.

This function has needed a small rewritting of the Rccp function
edgePlane from Morpho (here named edgePlane2) to have an additional
output that sortCurveMesh uses. Because edgePlane is called by
meshPlaneIntersect through edgePlaneIntersect, new versions of those 2
functions are proposed here (meshPlaneIntersect2 and
edgePlaneIntersect2), and were added to the utils_3D file. And because
meshPlaneIntersect was called within DrawOrthoplanes and selectPlanes
from our package (in the file select_planes_3D), some changes were
needed in those functions.

The modification in select_points_3D only concerns a plot correction of
intersection (now the points coordinates can contain NA values to
separate several isoltated outlines, that the old version of the
function didn't support).

Some lines were added in description and namespace files to ensure the
correct execution of the new plane functions.