Fix the approximated intersection plotted in the zoomed full mesh
Previously, when a decimation was made, and plane intersections were
required, those intersections were computed on the decimated mesh, and
directly projected onto the zoomed full mesh during landmark
positioning.
Now, once the intersections are computed on the decimated mesh, the
infos on the planes fixed by user on the decimated mesh are retrieved to
compute the intersection of those planes with the full mesh.
NOTE that an approximation still exist because those planes are
determined on the decimated mesh and not on the full mesh (centroid and
major planes of both meshes are not expected to be exactly the same, in
particular whe the decimation is high). It could be possible to set the
interactive determination of planes during user interaction on the
decimated mesh with the major planes determined on the full mesh. But an
approximation will still exist in that the relative orientation of major
planes from full mesh with the decimated mesh fixed by user won't be
exactly the same that if the interaction was done with the full mesh.

To do so, major changes occur in the DrawOrthoplanes() function from
selct_planes_3D. It was slightly generalized to deal with 2 cases:
- either the intersections among planes and mesh are determined and
plotted during a user interactive (as before)
- or the intersections are simply computed (with no plot) when plane
infos are input in the function.
Those plane infos are now output from the SelectPlanes function (from
select_planes_3D) through vPlanes.
In the main function DigitFixed (from digit_fixed_landmarks.r), there
are now 2 calls of the DrawoOthoplanes function:
- the first one (unchanged) to interactively compute and plot the
intersections with the decimated mesh
- the second one to compute the intersections with the full mesh (for
later use in SetPtZoom).