Give choice for major plane plotting and intersection
The previous version automatically plot the 3 major planes of the mesh
(when asked).
The change here consist in giving the choice to user of which plane(s)
is/are plotted (and which intersection(s) is/are computed).

The orthoplane argument in the DigitFixed function (from
digit_fixed_landmarks) was replaced by the argument drawPCplane which
can set to:
- FALSE -> no plane is plotted
- TRUE -> the 3 major planes are plotted
- any combination within the character vector
c("pc1-pc2","pc1-pc3","pc2-pc3") to plot either one or two planes
A check of this argument in this sense was added at the begining of the
function, to set idxPlanes . The calls of DrawOrthoplanes and SetPtZoom
now require this new variable (explaining the changes in those 2
functions, mainly with the replacement of the prefious for loop within
1:3 to the new ones within idxPlanes).