Add the option to have a single device divided into 2 synchronized subscenes
A new argument nbWin in the main function DigitFixed was added to
manage:
- either two separate windows during the digitization process (as
before) with nbWin=2
- or a unique window divided into 2 subscenes with nbWin=1

In the second setting, each translation and zoom done in one mesh will
affect automatically the second one.

The changes added in DigitFixed (digit_fixed_landmarks), Place Pt,
rgl.select2, setPtZoom (select_points_3D) and plot.landmark (utils_3D)
concern the mangement of landmarks plotting on the full and decimated
mesh, which differ depending on the number of devices used.