2 bug corrections + 1 amelioration
1st bug in selectPlanes: modify the initialization of vPlanes (to planes
rather than to an empty list) in case where the preplaced planes aren't
rotated by user, to have the initial planes stored in vPlanes
(previously, it caused a bug in SetPtZoom).
The planes are also poped once the user have validated them to only keep
the intersection lines.

2nd bug in SetPtZoom: when the intersection lines where reported on the
zoomed mesh, the distance among the clicked point and the intersection
points were computed independently for each plane in the for loop,
generating an approximative correspondence in the extend of the zoomed
mesh and the one for the intersection lines. It was corrected here by
applying the same distance threshold for the mesh and the intersection
extends.
NOTE: a bug still exits here when the zoomed submesh is made of  a stack
of separated meshes. Only the closets mesh to the clicked point is kept
through the use of vcgIsolated(), but the intersection lines with the
whole stack (contained in the percDist sphere around the clicked point)
will be plotted. It results in some intersection lines floatting behind
the zoomed mesh.

Amelioration in SetPtZoom: the zone coresponding to the zoomed mesh is
plotted in the window of the decimated mesh through a point cloud, to
better visualize to what location the zoomed mesh corresonds in the
whole shape.