Enhancement of graphical options specification (+ consequent changes in the functions) & bug fixes
1) The main change concerns a redefinition of graphical options used by
DigitFixed and subsequent function to set them more easily. Othe
arguments of DigitFixed were renamed for coherence.
Now DigitFixed needs 7 arguments:
- specFull (unchanged)
- decim (unchanged)
- fixed (unchanged)
- idxFixed (previous name: index)
- templateCoord (previous name: templateFile)
- idxTemplate (previous name: idxPtsTemplate)
- GrOpt: New one!

GrOpt is a list of 7 lists storing all graphical parameters needed for
DigitFixed. Each sublist being a set of parameters for a particular
graphical aspect. The 7 graphical categories are:
- window options (number, size & positioning, synchronization)
- mesh options (color, alpha, kind of plot)
- major planes options (which of them plot, color, alpha)
- mesh/plane intersection options (kind of plot, color)
- landmark spheres options (radius, color, alpha)
- landmark label options (size, color, adjustment)
- zoom options (extent, make visible or not the pre-placed landmark on
the zoomed mesh, and the zoom extent on the decimated mesh)
(for details see the function setDigitFixedOptions() in utils.r)

GrOpt is easily settable with default and/or customized values with the
new function setDigitFixedOptions() in the new file utils.r. By default,
the argument GrOpt in DigitFixed is set in this way:
GrOpt=setDigitFixedOptions().

Most of the changes in the modified functions correspond to the
management of these new options.
Note that the zoomSeeLm=TRUE option in GrOpt has needed more
modifications than the other ones (notably in rgl.select2())

2) 2 modifications added just before and after the call of setPtZoom()
in DigitFixed().

- Before this call, the previous version sent the coordinates of the
landmark positionned on the decimated mesh into the full mesh. This
landmark wasn't visible, but was used for deliming the submesh extent.
Now, zoomSeeLm=TRUE make it visible, and in case of high decimation,
this landmark can be not placed on the full mesh surface (and floating
around). To correct this visual bug, but also to base computation of
mesh extent not on this landmark, but on a landmark being at the sruface
of the full mesh, the project() function was called just before the call
of setPtZoom to project the landmark defined on the decimated mesh onto
the full one.

- After this call, the previous version called the project() function to
project onto the decimated mesh the landmark positionned on the full
mesh. Actually, it was correct, but because this projection is not
linked to points defining the mesh and can occur everywhere on triangle
surfaces, a virtual bug appears when the landmark pre-placed on the full
mesh is not modified by user and is directly validated: the landmark on
the decimated mesh move to its projection position which is not the
point from the decimated mesh on which it was first defined though
SelectPoints3D. So, in this case, whereas the user doesn't modify
landmark on the full mesh, its representation on the decimated mesh
slightly moves... This problem occurs because the rationale behind the
PlacePt() function is to positionned the sphere landmark at a mesh point
location (and not elsewhere eg on edges or on triangle surfaces), which
is not the rationnale of the project function() (in true
Morpho::projRead()), which can locate point everywhere on the triangles.
Though this rationale induce more precise landmark location than our, it
was simpler here to adopt our for the purpose of the correction of this
visual bug than to modify the rationale behind PlacePt() (but maybe it
would be changed later...).
So after the call of SetPtZoom, the project() function is not called
anymore, and is just replaced by a search of the most proximal mesh
point from the decimated mesh to the landmark placed on the full mesh.
Note: this bug is not entirely corrected and can still occur if the
landmark is next to 2 points from the decimated mesh.