Adds verbose
The change here is about the addition of verbose during the mesh
digitization (and more precisely during the plane adjustment, the
decimation, and the landmark digitization) to inform user on the step
which is processing, and possibly to pending actions.

Two levels of verbose were defined: the first one concerns verbose for
functions from our package, and the second one for functions from Rvcg.
An argument 'verbose' was added (taking the form of a 2-length logical
vector) to set this option, for the following functions (in
digit_fixed_landmarks.r and utils.r):
- digitMesh.mesh3d
- digitMesh.character
- decimMesh.character
and similar arguments (quiet and silent to respect the convention of the
associated functions from other packages) in:
- decimMesh.mesh3d
- read.tps

Corrections in the previous functions, and in some functions in
select_planes_3D.r, correposnd to the displaying of this verbose, as
well as the modification of the associated help to descibe this new
argument.