2 bug fixes
The first bug concerns the magnetism of landmark when using a single
window (divided into 2 scenes), occuring in PlacePt function (from
select_points_3D). In the previous version, the mesh vertices were
translated to width/2 to correspond to the x and y for the clicked
point. But it causes that the virtual user point assessed was
systematically deported toward the left, making wrong the magnetism of
the landmark (and possibly causing c++ crashes). Now this is the clicked
point which is translated toward the right (with an amount of width/2)
in order to correspond to the mesh vertices (kepts unchanged).

The second correction is the reduction of the array storing landmark
coordinates in digitMesh.character (utils.r), so that its third
dimension is now reduced if the user leaves the loop before to have
digitized all the mesh in the folder.