Fix bug on the names of meshes
Fix a bug on the output names of meshes (notably when processing a
single mesh object or name).

In the function digitMesh.mesh3d() from digit_fixed_landmarks.r the
correction concerned the moving of the line with deparse(subtitute(...))
which extracts the attribute coding for the mesh object name. Because
this line must be run before any modification af the mesh object which
can delete its attribute, it's necessary to move it before the mesh
cleaning steps (vcgUpdateNormals() & vcgClean()).
In addition, a new argument is added to deal with the call of this
function from digitMesh.character(), for which mesh name can't be
directly extracted from its attribute (help file is modified in
accordance).

In the function digitMesh.character(), the corrections concern the
storing of the results from the calling of digitMesh.mesh3d() (inluding
the extraction of mesh name from the ply filename), in order to
correctly extract the attribute name of the mesh, and to reaffect them
in the dimnames of the resulting array A.

Other corrections:
- some additions in namespace to export edgePlaneIntersect2() and to set
as dynamic the library digit3DLand (it was previulsy done, but it has
disapeared at some point)
- fix the default value of the sphere radius for landmarks in the help
to be in accordance with the arguments of setGraphicOptions().