Fix bug: NA values in TPS file
Fix some bugs in the function digitMesh.character() from utils.r in the
defintions of full.files & deci.files in the case where a folder with
meshes was partly processed during a previous digitization session.
Ultimately, the bugs impact the values of the TPS file which mainly
contained NA values.